Development, testing, and benchmarking of HashPartitionedReadWriteStorage (#5)
* Created HashPartitionedReadWriteStorage

Also revised benchmarking to include multiple time points

* Revised HashPartitionedReadWriteStorage, benchmarking, and testing

+++ src/main/java/dbf0/disk_key_value/readwrite/ReadWriteStorage.java
@@ -6,7 +6,7 @@ import javax.annotation.Nullable;
 import java.io.IOException;
 
 public interface ReadWriteStorage<K, V> {
-  int size() throws IOException;
+  long size() throws IOException;
 
   void put(@NotNull K key, @NotNull V value) throws IOException;
